Change bug fix to parser class instead of command class
