+++ src/main/java/duke/logic/commands/AddCommand.java
@@ -1,14 +1,10 @@
 package duke.logic.commands;
 
-import duke.commons.exceptions.ChronologyBeforePresentException;
-import duke.commons.exceptions.ChronologyInconsistentException;
 import duke.commons.exceptions.DuplicateTaskException;
 import duke.commons.exceptions.FileNotSavedException;
 import duke.logic.commands.results.CommandResultText;
 import duke.model.Event;
 import duke.model.Model;
-
-import java.time.LocalDateTime;
 
 /**
  * Adds a new event to users EventList.
@@ -32,32 +28,11 @@ public class AddCommand extends Command {
      * @param model The model object containing event list.
      * @throws FileNotSavedException If the data could not be saved.
      * @throws DuplicateTaskException If there is a duplicated event.
-     * @throws ChronologyBeforePresentException If the dates are before now.
-     * @throws ChronologyInconsistentException If the dates are inconsistent.
      */
     @Override
-    public CommandResultText execute(Model model) throws DuplicateTaskException, FileNotSavedException,
-            ChronologyInconsistentException, ChronologyBeforePresentException {
-        checkLogicalDate(event.getStartDate(), event.getEndDate());
+    public CommandResultText execute(Model model) throws DuplicateTaskException, FileNotSavedException {
         model.getEvents().add(event);
         model.save();
         return new CommandResultText(MESSAGE_ADDITION + event);
-    }
-
-    /**
-     * Checks if the dates provided is logical.
-     * Does nothing if it's logical.
-     * @param end The end date.
-     * @param start The start date.
-     * @throws ChronologyBeforePresentException If the dates are before now.
-     * @throws ChronologyInconsistentException If the dates are inconsistent.
-     */
-    private void checkLogicalDate(LocalDateTime start, LocalDateTime end) throws ChronologyBeforePresentException,
-            ChronologyInconsistentException {
-        if (start.isBefore(LocalDateTime.now()) || end.isBefore(LocalDateTime.now())) {
-            throw new ChronologyBeforePresentException();
-        } else if (end.isBefore(start)) {
-            throw new ChronologyInconsistentException();
-        }
     }
 }
